+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+FEATURES
+
+* Context now has access to the application-configured logger
+
 BREAKING CHANGES
 
 * types/rational now extends big.Rat
@@ -27,7 +31,6 @@ FEATURES:
   * Create genesis transactions with `gaiad init gen-tx` 
   * New genesis account keys are automatically added to the client keybase (introduce `--client-home` flag)
   * Initialize with genesis txs using `--gen-txs` flag
-* Context now has access to the application-configured logger
 
 BREAKING CHANGES